requirements.txt
requests==2.27.1
scikit-learn==1.0.2
SQLAlchemy==1.4.46
numpy==1.24.3
falcon==3.0.1
sseclient-py==1.7.2
gunicorn==20.1.0
matplotlib==3.5.1
scipy==1.8.0
